refactor(icon): address PR #740 review feedback
