+++ Sources/OvEditor/premake5.lua
@@ -18,7 +18,6 @@ project "OvEditor"
 		-- Dependencies
 		dependdir .. "glad/include",
 		dependdir .. "ImGui/include",
-		dependdir .. "stb_image/include",
 		dependdir .. "tinyxml2/include",
 		dependdir .. "tracy",
