+++ Sources/OvEditor/src/OvEditor/Panels/ProjectSettings.cpp
@@ -73,16 +73,16 @@ OvEditor::Panels::ProjectSettings::ProjectSettings(const std::string & p_title, 
 
 	{
 		/* Windowing settings */
-			auto& windowingRoot = CreateWidget<Layout::GroupCollapsable>("Windowing");
-			auto& columns = windowingRoot.CreateWidget<Layout::Columns<2>>();
-			columns.widths[0] = 125;
+		auto& windowingRoot = CreateWidget<Layout::GroupCollapsable>("Windowing");
+		auto& columns = windowingRoot.CreateWidget<Layout::Columns<2>>();
+		columns.widths[0] = 125;
 
-			GUIDrawer::DrawScalar<int>(columns, "Resolution X", GenerateGatherer<int>("x_resolution"), GenerateProvider<int>("x_resolution"), 1, 0, 10000);
-			GUIDrawer::DrawScalar<int>(columns, "Resolution Y", GenerateGatherer<int>("y_resolution"), GenerateProvider<int>("y_resolution"), 1, 0, 10000);
-			GUIDrawer::DrawBoolean(columns, "Fullscreen", GenerateGatherer<bool>("fullscreen"), GenerateProvider<bool>("fullscreen"));
-			GUIDrawer::DrawString(columns, "Executable name", GenerateGatherer<std::string>("executable_name"), GenerateProvider<std::string>("executable_name"));
-			GUIDrawer::DrawAsset(columns, "Window Icon", GenerateGatherer<std::string>("window_icon"), GenerateProvider<std::string>("window_icon"), OvTools::Utils::PathParser::EFileType::TEXTURE);
-		}
+		GUIDrawer::DrawScalar<int>(columns, "Resolution X", GenerateGatherer<int>("x_resolution"), GenerateProvider<int>("x_resolution"), 1, 0, 10000);
+		GUIDrawer::DrawScalar<int>(columns, "Resolution Y", GenerateGatherer<int>("y_resolution"), GenerateProvider<int>("y_resolution"), 1, 0, 10000);
+		GUIDrawer::DrawBoolean(columns, "Fullscreen", GenerateGatherer<bool>("fullscreen"), GenerateProvider<bool>("fullscreen"));
+		GUIDrawer::DrawString(columns, "Executable name", GenerateGatherer<std::string>("executable_name"), GenerateProvider<std::string>("executable_name"));
+		GUIDrawer::DrawAsset(columns, "Window Icon", GenerateGatherer<std::string>("window_icon"), GenerateProvider<std::string>("window_icon"), OvTools::Utils::PathParser::EFileType::TEXTURE);
+	}
 
 	{
 		/* Rendering settings */
